Add new secure fields from 10.2.7 and 11.0.0
This adds support for the AllowLoadGameType and UseSecureEnvironment
fields. The GuardedAddOn field was renamed to LoadFirst.

I've omitted support for the SuppressLocalTableRef and
AllowLoadByGameMode fields as these were abberations that only existed
for the duration of 10.2.6 and and 10.2.7 specifically; both are absent
as of 11.0.0 and with the current PTR clients for both Classic clients
are completely removed across all flavors.